done and will deploy this afternoon

+++ dojoreads_app/views.py
@@ -3,11 +3,15 @@ from dojoreads_app.models import *
 from django.shortcuts import render, redirect
 
 def books(request):
+    reviews = Review.objects.all().order_by("-updated_at")[0:3]
+    for review in reviews:
+        review.rating = range(review.rating)
+
     context = {
         "user_info": User.objects.get(id=request.session["userid"]),
         "all_books": Book.objects.all(),
-        "recent_books": Book.objects.order_by("-created_at")[:3],
-        "all_reviews": Review.objects.all()
+        "recent_books": Book.objects.order_by("-updated_at")[0:3],
+        "all_reviews": reviews
     }
     return render(request, "books.html", context)
 
@@ -48,9 +52,13 @@ def new_book_review(request):
         return render(request, "add-book-review.html", context)
 
 def book_info(request, book_id):
-    logged_user = User.objects.get(id=request.session["userid"])
-    selected_book = Book.objects.get(id=book_id)
+    book_exists_check = Book.objects.filter(id=book_id)
 
+    #redirect if book id doesn't exist
+    if not book_exists_check:
+        return redirect('/books')
+
+    selected_book = book_exists_check[0]
     context = {
         "user_info": User.objects.get(id=request.session["userid"]),
         "book_info": selected_book,
@@ -84,7 +92,12 @@ def user_profile(request, user_id):
     return render(request, 'user-profile.html', context)
 
 def delete_review(request, review_id):
-    review_to_delete = Review.objects.get(id=review_id)
-    book_id = review_to_delete.book_id.id
-    review_to_delete.delete()
+    review_to_delete = Review.objects.filter(id=review_id)
+    if not review_to_delete:
+        return redirect('/books')
+
+    if request.session["userid"] == review_to_delete[0].user_id.id:
+        book_id = review_to_delete[0].book_id.id
+        review_to_delete.delete()
+
     return redirect(f'/books/{book_id}')
